api/requirements1.txt
numpy==2.0.0
scipy==1.13.1
torch==2.3.1
IPython==8.18.1
pydub==0.25.1
fastapi==0.111.0
loguru==0.7.2
requests==2.32.3
ffmpeg-downloader==0.3.0
pickledb==0.9.2
python-snappy==0.7.2